opphouse/requirements.txt

django==2.1.2
gunicorn==19.9.0
djangorestframework==3.8.2
Pillow>=4.2.0
django-cms>=3.5.2